Revised Reshuffle Verification
Revised the Cypress commands for verifying the function that reshuffles the main stack of a deck without the drawn cards to be less dependent on conditional testing. Also added verification to check that the drawn cards are not in the main stack. Plus the test script for the positive test cases is more organized.