Set Up Custom Website and Cucumber
Set up the custom website to run in a local Node.js server. Also made Cucumber feature files and step definitions to test the website's content to the rough drafts of the homepage and the calculator page. Plus, the GitHub Actions workflow is updated to include the test report in the website when it's deployed.