Added Tests for Background Color Changes
Added tests to check that the background color of some elements changes when the mouse hovers over them. Also fixed an issue where the links to the homepage would not work if this website is published in GitHub Pages. Plus the default browser is Electron to allow the cypress-real-events plugin to work.